+++ Tests/RulesTests+Redundancy.swift
@@ -773,6 +773,11 @@ class RedundancyTests: RulesTests {
         testFormatting(for: input, rule: FormatRules.redundantInit)
     }
 
+    func testDontRemoveInitWhenUsedOnPossibleMetatype() {
+        let input = "let something = Foo.bar.init()"
+        testFormatting(for: input, rule: FormatRules.redundantInit)
+    }
+
     func testDontRemoveInitWithExplicitSignature() {
         let input = "[String.self].map(Foo.init(bar:))"
         testFormatting(for: input, rule: FormatRules.redundantInit)
@@ -863,6 +868,77 @@ class RedundancyTests: RulesTests {
         }
         """
         testFormatting(for: input, rule: FormatRules.redundantInit, exclude: ["indent"])
+    }
+
+    func testRemoveInitAfterCollectionLiterals() {
+        let input = """
+        let array = [String].init()
+        let arrayElement = [String].Element.init()
+        let nestedArray = [[String]].init()
+        let tupleArray = [(key: String, value: Int)].init()
+        let dictionary = [String: Int].init()
+        """
+        let output = """
+        let array = [String]()
+        let arrayElement = [String].Element()
+        let nestedArray = [[String]]()
+        let tupleArray = [(key: String, value: Int)]()
+        let dictionary = [String: Int]()
+        """
+        testFormatting(for: input, output, rule: FormatRules.redundantInit)
+    }
+
+    func testPreservesInitAfterTypeOfCall() {
+        let input = """
+        type(of: oldViewController).init()
+        """
+
+        testFormatting(for: input, rule: FormatRules.redundantInit)
+    }
+
+    func testRemoveInitAfterOptionalType() {
+        let input = """
+        let someOptional = String?.init("Foo")
+        // (String!.init("Foo") isn't valid Swift code, so we don't test for it)
+        """
+        let output = """
+        let someOptional = String?("Foo")
+        // (String!.init("Foo") isn't valid Swift code, so we don't test for it)
+        """
+
+        testFormatting(for: input, output, rule: FormatRules.redundantInit)
+    }
+
+    func testPreservesTryBeforeInit() {
+        let input = """
+        let throwing: Foo = try .init()
+        let throwingOptional1: Foo = try? .init()
+        let throwingOptional2: Foo = try! .init()
+        """
+
+        testFormatting(for: input, rule: FormatRules.redundantInit)
+    }
+
+    func testRemoveInitAfterGenericType() {
+        let input = """
+        let array = Array<String>.init()
+        let dictionary = Dictionary<String, Int>.init()
+        let atomicDictionary = Atomic<[String: Int]>.init()
+        """
+        let output = """
+        let array = Array<String>()
+        let dictionary = Dictionary<String, Int>()
+        let atomicDictionary = Atomic<[String: Int]>()
+        """
+
+        testFormatting(for: input, output, rule: FormatRules.redundantInit, exclude: ["typeSugar"])
+    }
+
+    func testPreserveNonRedundantInitInTernaryOperator() {
+        let input = """
+        let bar: Bar = (foo.isBar && bar.isBaaz) ? .init() : nil
+        """
+        testFormatting(for: input, rule: FormatRules.redundantInit)
     }
 
     // MARK: - redundantLetError
@@ -1661,6 +1737,18 @@ class RedundancyTests: RulesTests {
         let options = FormatOptions(redundantType: .explicit)
         testFormatting(for: input, rule: FormatRules.redundantType,
                        options: options)
+    }
+
+    func testRedundantTypeInModelClassNotStripped() {
+        // See: https://github.com/nicklockwood/SwiftFormat/issues/1649
+        let input = """
+        @Model
+        class FooBar {
+            var created: Date = Date.now
+        }
+        """
+        let options = FormatOptions(redundantType: .explicit)
+        testFormatting(for: input, rule: FormatRules.redundantType, options: options)
     }
 
     // --redundanttype infer-locals-only